requirements.txt
django
djangorestframework
pandas
openpyxl
faker
django-cors-headers